requirements.txt
bottle==0.12.9
meinheld==0.5.8
requests==2.11.1
pycrypto==2.6.1
